change type to uint

+++ 64bit_integer/include/lib.h
@@ -1,2 +1,4 @@
-unsigned long long *init_deck();
-void print_cards(unsigned long long cards);
+#include <stdint.h>
+
+uint64_t *init_deck();
+void print_cards(uint64_t cards);
